requirements.txt
# Core Python Libraries
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.8.2

# LangGraph and LangChain Components
langgraph==0.0.33
langchain==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
openai==1.23.6
langchain_chroma==0.2.4

# Streamlit Frontend
streamlit==1.46.0

# Coinbase CDP SDK
cdp-sdk==1.15.0

# Vector Store
chromadb==0.4.24

# Async Support
nest_asyncio==1.6.0

# HTTP Client
httpx==0.27.0

# JSON and Data Processing
orjson==3.10.7

# Environment and Configuration
python-multipart==0.0.20

# Optional: For better performance
uvloop==0.19.0